Commit temporary project architecture
